requirements.txt
alembic==1.6.5
amqp==5.0.6
autopep8==1.5.7
Babel==2.9.1
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
cachelib==0.2.0
celery==5.1.2
cffi==1.14.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.0.9
distlib==0.3.2
dnspython==2.1.0
email-validator==1.1.3
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
Flask-Babel==2.0.0
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-GraphQL==2.0.1
Flask-JWT==0.3.2
Flask-JWT-Extended==4.2.3
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==3.0.1
Flask-Moment==1.0.2
Flask-PageDown==0.3.0
flask-react==0.0.5
Flask-REST==1.3
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
graphql-core==2.3.2
graphql-server-core==1.2.0
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
infinity==1.5
intervals==0.9.2
is-disposable-email==1.0.0
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
limits==1.5.1
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
pkg_resources==0.0.0
platformdirs==2.0.2
promise==2.3
prompt-toolkit==3.0.19
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
redis==3.5.3
rope==0.19.0
rq==1.9.0
Rx==1.6.1
six==1.16.0
SQLAlchemy==1.4.21
toml==0.10.2
validators==0.18.2
vine==5.0.0
virtualenv==20.6.0
wcwidth==0.2.5
Werkzeug==2.0.1
WTForms==2.3.3
WTForms-Components==0.10.5
wtforms-validators==1.0.0
